weekly statistic works - wip
