+++ src/app/domains/statistics/weekly-statistics.component.html
@@ -1,0 +1,3 @@
+<div #statisticsContainer>
+  <svg id="statistics-graph" width="100%" attr.height="{{ svgComponentHeight }}px"></svg>
+</div>
